add failure message for untyped args to main
We don't try to infer the base types of the arguments of main, so the
user must annotate them. So far, if these annotations were not present,
we did not add them to the 'pre_vars', i.e., the predefined vars, which
eventually resulted in a cryptic "key not found" error. This commit
makes drift fail explicitly in this situation and warns about the
possible problem.